interesting idea
use the chassisspeeds methods to convert the field relative robot speed to a speed relative to the robot IF it were completely autoaligned to the speaker, then use that speed to modify the omega value in autoalign to account for robot movement. Implementation currently only works if robot is moving along radial/in a semicircle around speaker.